langraph-agent/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.45
python-dotenv==1.0.1
pydantic==2.9.2
PyMuPDF==1.24.13
Pillow==11.0.0
httpx==0.27.2

